requirements.txt
Django==3.1.3
gunicorn==20.0.4
python-dotenv==0.15.0
requests==2.22.0
whitenoise==5.2.0
python-socketio==5.0.4
eventlet==1.14.0
aiohttp==3.7.3
uvicorn[standard]==7.1.2